+++ R/conn.open_gdal.R
@@ -11,39 +11,44 @@
    a <- try(rgdal::GDALinfo(fname,returnStats=FALSE,returnRAT=FALSE
                 ,returnColorTable=TRUE,returnCategoryNames=TRUE),silent=TRUE)
    if (inherits(a,"try-error")) {
-      if ((TRUE)||(!.isPackageInUse()))
-         cat(geterrmessage())
-     # * using R version 3.4.0 Patched (2017-05-16 r72684)
-     # * using platform: x86_64-pc-linux-gnu (64-bit)
-     # > ### Name: open_gdal
-     # > ### Title: Open GDAL file
-     # > ### Aliases: open_gdal
-     # > ### Keywords: connection
-     # > 
-     # > ### ** Examples
-     # > 
-     # > session_grid(NULL)
-     # > fname1 <- system.file("pictures/cea.tif",package="rgdal")
-     # > message(fname1)
-     # /tmp/RtmpdrKt9J/RLIBS_310d3548ed74/rgdal/pictures/cea.tif
-     # > a1 <- open_gdal(fname1)
-     # > print(a1)
-     # NULL
-     # > print(a1[])
-     # NULL
-     # > close(a1)
-     # Error in UseMethod("close") : 
-     #   no applicable method for 'close' applied to an object of class "NULL"
-     # Calls: close
-     ## 20170529 patch for failure with 'rgdal' at r-forge
-      if (.Platform$OS.type=="unix"){
-         message(paste("Unable to open GDAL file. Failure for"
-                      ,"R-forge buildig machine (Unix OS) since May 2017"))
-         if ((!.lgrep("\\.(rds)$",fname))&&(file.exists(fname))) {## 20170529
-            return(ursa_new())
-         }
-      } 
-      return(NULL) 
+      fname <- normalizePath(fname)
+      a <- try(rgdal::GDALinfo(fname,returnStats=FALSE,returnRAT=FALSE
+                   ,returnColorTable=TRUE,returnCategoryNames=TRUE),silent=TRUE)
+      if (inherits(a,"try-error")) {
+         if ((TRUE)||(!.isPackageInUse()))
+            cat(geterrmessage())
+        # * using R version 3.4.0 Patched (2017-05-16 r72684)
+        # * using platform: x86_64-pc-linux-gnu (64-bit)
+        # > ### Name: open_gdal
+        # > ### Title: Open GDAL file
+        # > ### Aliases: open_gdal
+        # > ### Keywords: connection
+        # > 
+        # > ### ** Examples
+        # > 
+        # > session_grid(NULL)
+        # > fname1 <- system.file("pictures/cea.tif",package="rgdal")
+        # > message(fname1)
+        # /tmp/RtmpdrKt9J/RLIBS_310d3548ed74/rgdal/pictures/cea.tif
+        # > a1 <- open_gdal(fname1)
+        # > print(a1)
+        # NULL
+        # > print(a1[])
+        # NULL
+        # > close(a1)
+        # Error in UseMethod("close") : 
+        #   no applicable method for 'close' applied to an object of class "NULL"
+        # Calls: close
+        ## 20170529 patch for failure with 'rgdal' at r-forge
+         if (.Platform$OS.type=="unix"){
+            message(paste("Unable to open GDAL file. Failure for"
+                         ,"R-forge buildig machine (Unix OS) since May 2017"))
+            if ((!.lgrep("\\.(rds)$",fname))&&(file.exists(fname))) {## 20170529
+               return(ursa_new())
+            }
+         } 
+         return(NULL) 
+      }
    }
    a1 <- as.numeric(a)
    g1 <- regrid()