+++ R/xxx.panel_legend.R
@@ -10,8 +10,11 @@
                            ##~ ,seg.len=2)
 '.shape_legend<-' <- function(obj,items,value) {
    res <- unlist(lapply(items,function(x) x[[value]]))
-   if (!is.null(res))
+   if (!is.null(res)) {
+      str(value)
+      str(res)
       obj[[value]] <- res
+   }
    obj
 }
 'panel_legend' <- function(items,...)
@@ -20,6 +23,7 @@
       return(NULL)
    if (missing(items))
       items <- getOption("ursaPngLegend")
+   items <- items[!sapply(items,is.null)]
    if (!length(items))
       return(invisible(NULL))
    arglist <- as.list(args(legend))
@@ -33,11 +37,17 @@
    if (is.language(arglist[["box.lwd"]]))
       arglist[["box.lwd"]] <- 0.1
    if (is.language(arglist[["pt.cex"]]))
-      arglist[["pt.cex"]] <- 1#arglist[["cex"]]
+      arglist[["pt.cex"]] <- 1 #arglist[["cex"]]
    if (is.language(arglist[["pt.lwd"]]))
       arglist[["pt.lwd"]] <- arglist[["lwd"]]
    if (!nchar(as.character(arglist[["x"]])))
       arglist[["x"]] <- "bottomright"
+   if (all(is.na(arglist[["lwd"]])))
+      arglist[["lwd"]] <- NULL
+   if (all(is.na(arglist[["lty"]])))
+      arglist[["lty"]] <- NULL
+   if (all(is.na(arglist[["cex"]])))
+      arglist[["cex"]] <- NULL
    arglist[["merge"]] <- FALSE ## '$merge : language do.lines && has.pch'
    for (a in names(arglist)) {
       if (is.language(arglist[[a]])) {